Feat - auth to listing service (#4)
* add auth to listing service

* nits

+++ application.py
@@ -7,6 +7,7 @@ import re
 from models.Listing import ListingQueryModel
 import boto3
 import os
+from flask_jwt_extended import (JWTManager, create_access_token, create_refresh_token, get_jwt_identity, jwt_required)
 
 camel_pat = re.compile(r'([A-Z])')
 under_pat = re.compile(r'_([a-z])')
@@ -26,11 +27,13 @@ def convert_json(d, convert):
 # Create the Flask application object.
 application = app = Flask(__name__)
 SNS_CLIENT = boto3.client('sns', region_name='us-east-1')
-
 CORS(app)
-def send_new_listing_notif(lid):
+app.config['SECRET_KEY'] = os.getenv("SECRET_KEY")
+jwt = JWTManager(application)
+
+def send_new_listing_notif(listing):
     # sends a notification to SNS on new listing created
-    message = "New Listing {} just came up! Check it out!".format(lid)
+    message = "New Listing {} just came up! Check it out! Only with ${}".format(listing.listing_id, listing.price)
     response = SNS_CLIENT.publish(
         TargetArn=os.environ.get("SNS_ARN"),
         Message=json.dumps({'default': message}),
@@ -56,6 +59,7 @@ def get_health():
 
 
 @app.route("/api/listings", methods=["GET"])
+@jwt_required()
 def get_all_listings():
     try:
         with ListingQueryModel() as lqm:
@@ -73,11 +77,15 @@ def get_all_listings():
         return rsp
 
 @app.route("/api/listings", methods=["POST"])
+@jwt_required()
 def create_listing():
     try:
         listing_info = convert_json(request.get_json(), camel_to_underscore)
         with ListingQueryModel() as lqm:
             listing = lqm.add_listing(listing_info)
+            # FIXME: should not send in development
+            # send SNS
+            send_new_listing_notif(listing)
             rsp = Response(
                 json.dumps(convert_json(serialize(listing),underscore_to_camel)), 
                 status=200,
@@ -91,12 +99,24 @@ def create_listing():
 
 
 @app.route("/api/listing/<lid>", methods=["GET", "PUT", "DELETE"])
+@jwt_required()
 def listing_info_id(lid):
     def get_listing_by_id(lid):
         with ListingQueryModel() as lqm:
             return lqm.get_listing_by_id(lid)
-
     try:
+        current_uid = get_jwt_identity()
+        # listing = get_listing_by_id(lid)
+        # if listing:
+        #     listing_uid = listing.author_user_id
+        #     current_uid = get_jwt_identity()
+        #     if listing.author_user_id != current_uid:
+        #         return Response("Inconsistent user", status=401,
+        #                     content_type="text/plain")
+        # else:
+        #     return Response("listing not found", status=404,
+        #                        content_type="text/plain")
+        
         if request.method == "GET":
             listing = get_listing_by_id(lid)
             if listing:
@@ -114,11 +134,10 @@ def listing_info_id(lid):
             listing_info = convert_json(request.get_json(), camel_to_underscore)
             with ListingQueryModel() as lqm:
                 lqm.update_listing_by_id(lid, listing_info)
-                
-                # FIXME: should not send in development
-                # send SNS
-                send_new_listing_notif(lid)
                 listing = get_listing_by_id(lid)
+                if listing.author_user_id != current_uid:
+                    return Response("Inconsistent user", status=401,
+                                content_type="text/plain")
                 rsp = Response(
                     json.dumps(convert_json(serialize(listing),underscore_to_camel)), 
                     status=200,
@@ -128,6 +147,9 @@ def listing_info_id(lid):
         elif request.method == "DELETE":
             listing = get_listing_by_id(lid)
             if listing:
+                if listing.author_user_id != current_uid:
+                    return Response("Inconsistent user", status=401,
+                                content_type="text/plain")
                 with ListingQueryModel() as lqm:
                     lqm.delete_listing_by_id(lid)
                     rsp = Response("listing with lid {} is successfully deleted.".format(lid), status=200,
@@ -187,4 +209,4 @@ def serialize(listings):
 
 
 if __name__ == "__main__":
-    app.run(host="127.0.0.1", port=7001, debug=True)
+    app.run(host="localhost", port=7001, debug=True)
